// src/app/model/CartItems.model.ts
export class CartItems{
    id: number | undefined;
    productCode: string |undefined;
    productName: string |undefined;
    userId: number | undefined;
    productId: number | undefined;
    price: number | undefined;
    quantity?: number;
    totalPrice: number | undefined;
    mainImage?: string | undefined;
}